Count sum of КБЖУ

+++ app/src/main/java/com/example/caloriecalc/data/NutrientSummary.kt
@@ -1,8 +1,8 @@
 package com.example.caloriecalc.data
 
 data class NutrientSummary(
-    val calories: Int,
-    val protein: Float,
-    val fats: Float,
-    val carbs: Float
+    val calories: Double,
+    val protein: Double,
+    val fats: Double,
+    val carbs: Double
 )
